🎨 FinishedTasks 컴포넌트 분리
- HistoryPage에 모두 작성되었던 내용을 FinishedTasks컴포넌트를 생성하여 분리함

- HistoryPage에는 페이지 전체를 나타내는 레이아웃이 남도록 하고, 여기에 페이지를 구성하는 컴포넌트를 추가
- scss 모듈을 가져오는 이름을 style -> styles로 변경
