+++ src/routes/historyPage/HistoryPage.module.scss
@@ -32,27 +32,37 @@
 
   .tasks {
     width: 100%;
-  }
 
-  .task {
-    @include flexbox.flexbox(center, center);
-    width: 100%;
-    height: 50px;
-    background-color: colors.$WHITE;
-    border-radius: 15px;
-    box-shadow: 0 10px 20px 0 rgba(colors.$BLACK, 2%);
+    .task {
+      @include flexbox.flexbox(center, center);
+      width: 100%;
+      height: 50px;
+      background-color: colors.$WHITE;
+      border-radius: 15px;
+      box-shadow: 0 10px 20px 0 rgba(colors.$BLACK, 2%);
 
-    + .task {
-      margin-top: 10px;
-    }
+      + .task {
+        margin-top: 10px;
+      }
 
-    .title {
-      margin: 0 40px;
-      overflow: hidden;
-      line-height: normal;
-      color: colors.$TASK_ITEM_TITLE;
-      text-overflow: ellipsis;
-      white-space: nowrap;
+      .title {
+        width: 75%;
+        margin: 0 0 0 40px;
+        overflow: hidden;
+        line-height: normal;
+        color: colors.$TASK_ITEM_TITLE;
+        text-overflow: ellipsis;
+        white-space: nowrap;
+      }
+
+      .deleteIcon {
+        display: none;
+        cursor: pointer;
+      }
+
+      &:hover .deleteIcon {
+        display: block;
+      }
     }
   }
 }
